[#111]feat:팀 갤러리 인증마크, 네비게이션바 빨간 점 추가
