--- src/components/Meeting22/NavigationBar/Styles.js
@@ -1,44 +1,0 @@
-import styled from 'styled-components';
-
-export const NavBarContainer = styled.nav`
-  position: sticky;
-  left: 0;
-  bottom: 0;
-  width: 100%;
-  height: 12%;
-  background-color: #ffffff;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  border-top: 1px solid #f0f0f0;
-`;
-
-export const NavItem = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  height: 100%;
-  font-size: 12px;
-  color: #151515;
-  text-align: center;
-  cursor: pointer;
-  padding: 5px 0;
-`;
-
-export const IconWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const RedDot = styled.div`
-  position: absolute;
-  top: 0px;   
-  left: 0px;  
-  width: 8px;
-  height: 8px;
-  background-color: red;
-  border-radius: 50%;
-  z-index: 10;
-`;
